Update Group Page
Made the code and the page a little more comprehensible. Added a short description of what the team is working on and a list of members with links to our individual pages for those who uploaded theirs already. As long as the files are uploaded in the same structure as the repo it should "just work".